fix(build): Complete build system fixes for advanced consensus features
Build System:
- Add evo/*.cpp, bls/*.cpp, llmq/*.cpp, assets/atomicswap.cpp to Makefile.am

Header Fixes:
- Add mutable keyword to CEvoDB::cs for const method locking
- Add primitives/transaction.h include to deterministicmns.h for COutPoint
- Add deserializing constructor to CDeterministicMN
- Add pubkey.h include to bls/bls.h for CKeyID type
- Add CBlock/CBlockIndex forward declarations to providertx.h
- Add specialtx.h include to deterministicmns.cpp and providertx.cpp
- Add accessor methods for CDeterministicMNList private members
- Add required includes to atomicswap.cpp (chainparams.h, net.h, policy/fees.h)

API Corrections:
- Replace OPENSSL_cleanse with memory_cleanse in bls/bls.cpp
- Replace Params() with GetParams() in atomicswap.cpp
- Replace LookupBlockIndex() with mapBlockIndex.find() in chainlocks.cpp
- Fix AcceptToMemoryPool signature (add fMissingInputs, remove bad fee param)
- Replace RelayTransaction with manual CInv broadcast
- Replace GetMinRelayFee() with minRelayTxFee.GetFee()
- Fix CDBIterator::GetKey() usage with proper key type
- Cast LLMQType enum to uint8_t for serialization
- Cast integer sums to int64_t for UniValue::pushKV
- Fix CRPCCommand structures (remove obsolete bool field)
- Fix TransactionSignatureCreator with CTransaction instead of CMutableTransaction

+++ src/llmq/instantsend.cpp
@@ -47,7 +47,7 @@ uint256 CInstantSendLock::GetRequestId() const
 uint256 CInstantSendLock::GetSignHash() const
 {
     CHashWriter hw(SER_GETHASH, PROTOCOL_VERSION);
-    hw << INSTANTSEND_QUORUM_TYPE;
+    hw << static_cast<uint8_t>(INSTANTSEND_QUORUM_TYPE);
     hw << quorumHash;
     hw << GetRequestId();
     hw << txid;
